fix 1

+++ src/main/java/ru/practicum/shareit/user/service/UserServiceImpl.java
@@ -4,7 +4,9 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import ru.practicum.shareit.exception.NotFoundException;
 import ru.practicum.shareit.user.model.User;
-import ru.practicum.shareit.user.dto.UserDto;
+import ru.practicum.shareit.user.dto.UserCreateRequest;
+import ru.practicum.shareit.user.dto.UserResponseDto;
+import ru.practicum.shareit.user.dto.UserUpdateRequest;
 import ru.practicum.shareit.user.dto.UserMapper;
 import ru.practicum.shareit.user.storage.InMemoryUserStorage;
 
@@ -17,44 +19,42 @@ public class UserServiceImpl implements UserService {
     private final InMemoryUserStorage userStorage;
 
     @Override
-    public UserDto createUser(UserDto userDto) {
-        User user = UserMapper.fromUserDto(userDto);
+    public UserResponseDto createUser(UserCreateRequest createRequest) {
+        User user = UserMapper.fromCreateRequest(createRequest);
         User createdUser = userStorage.create(user);
-        return UserMapper.toUserDto(createdUser);
+        return UserMapper.toResponseDto(createdUser);
     }
 
     @Override
-    public UserDto updateUser(Long userId, UserDto userDto) {
+    public UserResponseDto updateUser(Long userId, UserUpdateRequest updateRequest) {
         User existingUser = userStorage.findById(userId)
                 .orElseThrow(() -> new NotFoundException("Пользователь с id " + userId + " не найден"));
 
-        if (userDto.getName() != null) {
-            existingUser.setName(userDto.getName());
-        }
-        if (userDto.getEmail() != null) {
-            existingUser.setEmail(userDto.getEmail());
-        }
+        UserMapper.updateUserFromRequest(updateRequest, existingUser);
 
         User updatedUser = userStorage.update(existingUser);
-        return UserMapper.toUserDto(updatedUser);
+        return UserMapper.toResponseDto(updatedUser);
     }
 
     @Override
-    public UserDto getUserById(Long id) {
+    public UserResponseDto getUserById(Long id) {
         User user = userStorage.findById(id)
                 .orElseThrow(() -> new NotFoundException("Пользователь с id " + id + " не найден"));
-        return UserMapper.toUserDto(user);
+        return UserMapper.toResponseDto(user);
     }
 
     @Override
-    public List<UserDto> getAllUsers() {
+    public List<UserResponseDto> getAllUsers() {
         return userStorage.findAll().stream()
-                .map(UserMapper::toUserDto)
+                .map(UserMapper::toResponseDto)
                 .collect(Collectors.toList());
     }
 
     @Override
     public void deleteUser(Long id) {
+        if (!userStorage.findById(id).isPresent()) {
+            throw new NotFoundException("Пользователь с id " + id + " не найден");
+        }
         userStorage.delete(id);
     }
 }
